% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nbn_data.R
\name{download_nbn_data}
\alias{download_nbn_data}
\title{Download NBN data}
\usage{
download_nbn_data(request)
}
\arguments{
\item{request}{a request string}
}
\value{
an sf object
}
\description{
Download data from an nbn endpoint
}
\details{
This function handles steps to download, unzip and read shapefile data
}
