% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbn_occurence.R, R/get_occurence.R
\name{get_occurence}
\alias{get_occurence}
\title{Get Occurance}
\usage{
get_occurence(taxon, lat, lon, radius_km)

get_occurence(taxon, lat, lon, radius_km)
}
\arguments{
\item{taxon}{a string detailing the taxon name}

\item{lat}{the latitude coordinate}

\item{lon}{the longitude coordinate}

\item{radius_km}{the radius in kilometers within which to return records}
}
\description{
Get NBN Occurance Data

Get NBN Occurance Data
}
